+++ src/Controller/UserController.php
@@ -38,6 +38,10 @@ class UserController extends AbstractController
      */
     public function showUserAction(Request $request,$id)
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            return $this->redirectToRoute('app_login');
+        }
+
         if ($id) {
             $em = $this->getDoctrine()->getManager();
             /** @var User $user */
@@ -58,6 +62,10 @@ class UserController extends AbstractController
      */
     public function new(Request $request, UserPasswordEncoderInterface $passwordEncoder): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            return $this->redirectToRoute('app_login');
+        }
+
         $user = new User();
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
@@ -92,6 +100,10 @@ class UserController extends AbstractController
      */
     public function edit(Request $request, UserPasswordEncoderInterface $passwordEncoder, User $user): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            return $this->redirectToRoute('app_login');
+        }
+
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
 
@@ -123,6 +135,10 @@ class UserController extends AbstractController
      */
     public function delete(Request $request, User $user): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            return $this->redirectToRoute('app_login');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($user);
